+++ CHANGELOG.md
@@ -31,6 +31,7 @@
   compilation time. Lower levels may improve compile times but reduce run-time performance. It can be configured
   globally, or per-module using `wp.set_module_options({"optimization_level": #})`. Currently it only has an effect
   on GPU modules and only when using CUDA Toolkit 12.9 or newer ([GH-1084](https://github.com/NVIDIA/warp/issues/1084)).
+- Add `wp.tile_randi()` and `wp.tile_randf()` to support generating tiles of random numbers ([GH-1010](https://github.com/NVIDIA/warp/issues/1010))
 - Add `wp.mesh_query_ray_anyhit` for ray any-hit queries ([GH-1097](https://github.com/NVIDIA/warp/issues/1097)).
 - Add support for group-aware construction and queries for `wp.Mesh` ([GH-1097](https://github.com/NVIDIA/warp/issues/1097)):
   - New constructor argument: `groups` for per-face group IDs.